feat: Add play screen and selection logic
Implemented Play Screen only with the centered board
Implemented selection logic for pieces on the Play Screen
Refs: #1
