+++ Checkers/Board/BoardViewModel.swift
@@ -1,0 +1,5 @@
+import SwiftUI
+
+final class BoardViewModel: ObservableObject {
+  @Published var pieces = Pieces.start
+}
