+++ Checkers/CheckersApp.swift
@@ -2,9 +2,11 @@ import SwiftUI
 
 @main
 struct CheckersApp: App {
-    var body: some Scene {
-        WindowGroup {
-          HomeView()
-        }
+  var body: some Scene {
+    WindowGroup {
+      NavigationView {
+        HomeView()
+      }
     }
+  }
 }
